+++ end_point.py
@@ -4,31 +4,12 @@ from sqlquery import query_more
 from sqlquery import ins_ika
 
 
-class Ika():
-    '''one comment that includes:[Ika_id, poster, forward_Ika_id]'''
-    def __init__(self, ika_id, forward_ika, post_time, poster_id, poster_name='NaMe', comment="hello"):
-        '''
-        Ika_id is a unique id for every Ika.
-        forward is the Ika id that this Ika link to
-        poster id is the id belongs to poster
-        name and email can be null and will set to 'NaMe' and 'Null'
-        comment means content
-        '''
-        self.ika_id = ika_id
-        self.forward_ika = forward_ika
-        self.post_time = post_time
-        self.poster_id = poster_id
-        self.poster_name = poster_name
-        self.comment = comment
-
-
-
 def get_ika(ika_id):
     '''
     get the spec Ika by its id
     return a Ika
     '''
-    return sqlquery.query_by_id(ika_id)
+    return query_by_id(ika_id)
 
 
 def get_reply(fatherid, num_begin, num_end):
@@ -39,7 +20,7 @@ def get_reply(fatherid, num_begin, num_end):
     num_begin > 0
     return a [Ika, ...]
     '''
-    return sqlquery.query_more(fatherid, num_begin, num_end)
+    return query_more(fatherid, num_begin, num_end)
 
 
 def new_ika(forward_ika, poster_id, poster_name, comment):
@@ -47,7 +28,7 @@ def new_ika(forward_ika, poster_id, poster_name, comment):
     post a new ika
     return None
     '''
-    sqlquery.ins_ika(forward_ika, poster_id, poster_name, comment)
+    ins_ika(forward_ika, poster_id, poster_name, comment)
 
 
 def new_poster():
@@ -55,4 +36,4 @@ def new_poster():
     return a unique id for each user
     return int
     '''
-    return rando.getno()
+    return getno()
